chore(release.yml): update release workflow to create and upload a zip archive
The release workflow has been updated to create a zip archive of the repository and upload it as an artifact. This will allow users to easily download and use the latest release of the project. The archive is created using the `thedoctor0/zip-release` action, and the resulting zip file is named "release.zip". The `.git` directory and its contents are excluded from the archive. The `ncipollo/release-action` action is then used to upload the zip file as an artifact.